requirements.txt
Flask
gunicorn
scikit-learn
pandas
numpy
joblib
google-generativeai